blkif: Avoid use-after-free on BLKIF_OP_WRITE_BARRIER request
Function tapdisk_xenblkif_make_vbd_request will release
tapreq->msg and reinsert it in the reqs_free array if the
request is a BLKIF_OP_WRITE_BARRIER without any segment.
But tapdisk_xenblkif_queue_request will use it to check if
it must send the request to VBD layer.

This bug triggers a crash when DEBUG is defined in td-reqs.c
because DEBUG will poison the msg memory.
Without DEBUG, the code reads freed memory that still contains
unchanged data.

Signed-off-by: Anthoine Bourgeois <anthoine.bourgeois@vates.tech>