vbd: wake-up scheduler to force check ring
In the case of a write barrier, the ring is only explore until the
barrier lefting pending requests hereafter. The completion of a barrier
will trigger a ring check but if the select is already sleeping it will
not wake it up so it will have to wait the next wake up or timeout.

Signed-off-by: Anthoine Bourgeois <anthoine.bourgeois@vates.tech>